==> Notebooks/1_Scraping.ipynb <==
import pandas as pd
import numpy as np
import pickle
import getpass

import sys
sys.path.append("../")
import src.Resources_scrap as src

import warnings
warnings.filterwarnings('ignore')

Here i call my Resources folder and execute my imdb scraping function. Once it has scraped all the needed data it will store each genre in a pickle in our data/pickles folder.

for n in range(1,5):
    src.search(n)

Goodreads scraping

In this notebook, we will scrap books from Goodreads. To achieve this we need an account in there because goodreads doesn't let you navigate through various pages without it, that's why I have an Email and Password input here, as well my function receives a list of genres that it will automatically scrap.


Email = getpass.getpass()
Password = getpass.getpass()
genres= ["Action","Adventure","Comedy","Crime","Drama","Fantasy","History","Horror","Mystery","Romance","Sciencefiction","Superhero","Thriller"]

src.search(genres,Email,Password)

In this other function we will scrap a specific book list by giving a link to my function, and it will iterate through the webpage up to 5000 times, this can be changed to be higher inside the function.


Book_list = "https://www.goodreads.com/list/show/264.Books_That_Everyone_Should_Read_At_Least_Once"

src.search_list(Book_list,Email,Password)
