==> Notebooks/2_Cleaning.ipynb <==
import pandas as pd
import numpy as np
import pickle
import sys
sys.path.append("../")
import src.Resources as src

import warnings
warnings.filterwarnings('ignore')


In this notebook i will clean all the data from imdb. We start with this base csv called Movies.

main_df = pd.read_csv('../Data/Movies.csv', index_col= 0)  

We change the name of each of the collumns

main_df.columns = ["Movie_name","Movie_img","Movie_genre","Movie_description","Movie_rating","Casting","Movie_votes","Movie_directors","Movie_stars"]

We run two cleaning functions from Resources.py

main_df["Movie_rating"] = main_df["Movie_rating"].apply(lambda x: src.movie_rating_cleaner(x))
main_df["Movie_votes"] = main_df["Movie_votes"].apply(lambda x: src.movie_vote_cleaner(x))

And now it's time to run the main cleaning function for imdb, it also grabs all the .pkl we have in data and concat them together to make one DataFrame of them all.

clean_df = src.movie_cleaner(main_df)

Here we need to reset the index, drop duplicates and reset index again.

clean_df.reset_index(inplace=True)
clean_df.drop("index",axis=1,inplace=True)
clean_df.drop_duplicates(inplace=True)
clean_df.reset_index(inplace=True)
clean_df.drop("index",axis=1,inplace=True)

Here we divide our Directors and Stars by the character "|", this will be useful later on for my recommendation.

clean_df["Movie_stars"] = clean_df["Movie_stars"].apply(lambda x: src.split_casting(x))
clean_df["Movie_directors"] = clean_df["Movie_directors"].apply(lambda x: src.split_casting(x))

I don't need casting anymore so we will drop it.

clean_df.drop("Casting",axis=1,inplace=True)

We make sure that there aren't any null values, and then we save it as Movies_clean

clean_df["Movie_directors"]

clean_df.isnull().sum()

clean_df.to_csv("../Data/Movies_clean.csv")

Now we will clean the Goodreads scraped data

In this notebook I will clean the data scraped from Goodreads and as well put it all in one DataFrame.
First, we start with the DataFrame Books as a baseline.

book_df = pd.read_csv('../Data/Books.csv', index_col= 0)  

book_df.head(1)

This is a list of all the .pkl that we want to load to our DataFrame, i chose them all, if in a future you would scrap new books you need to put the name of the pkl in here aswell, keep in mind to remove Books from the name here.

Exemple: Books_Action.pkl for this list would be just Action.

genres= ["Action","Adventure","Comedy","Crime","Drama","Fantasy","History","Horror","Mystery","Romance","Sciencefiction","Superhero","Thriller","List"]

Here we run the cleaning function, this function aswell puts all the pickles together into one DataFrame.

book_df = src.book_cleaner(genres,book_df)

We reset the index, drop duplicates, reset index again and then we multiply the rating by 2 so its in the same scale as imdb.

book_df.reset_index(inplace=True)
book_df.drop("index",axis=1,inplace=True)
book_df.drop_duplicates(inplace=True)
book_df.reset_index(inplace=True)
book_df.drop("index",axis=1,inplace=True)
book_df["Book_rating"] = book_df["Book_rating"].apply(lambda x: src.book_rating_multiplier(x))

We make sure that there aren't any Nan values, in this case there is one so ill just replace it for: "No description".

book_df.isnull().sum()

book_df=book_df.fillna("No description")

This is the look of the DataFrame clean and with all of the new data we got from the pickles that were scraped with our goodreads selenium function.

And finally we save it as Books_clean

book_df.to_csv("../Data/Books_clean.csv")

Here we load Books_clean to simplify the genres.

book_df2 = pd.read_csv('../Data/Books_clean.csv', index_col= 0)  

This function uses a dictionary in my library.py to simplify all the genres of the books, then this is applied with a map.

book_df2['Book_genre'] = book_df['Book_genre'].apply(src.map_genre)

These are now the new genres for the books.

book_df2

We drop any Nan because that means that it doesn't coincide with any genre that i have for my movies, so I won't be able to do a good recommendation based in the genre.

book_df2.isna().sum()

book_df2.dropna(inplace=True)

book_df2

book_df2.to_csv("../Data/Books_clean_filtered.csv")

Now we do the cleaning to prepare the recomendation

movies = pd.read_csv("../Data/Movies_clean.csv",index_col=0)
books = pd.read_csv("../Data/Books_clean_filtered.csv",index_col=0)

We rename our columns

movies.columns = ["Title","Image","Genre","Description","Rating","Votes","Directors","Stars"]
books.columns = ["Author","Image","Description","Rating","Votes","Title","Genre"]

Drop unnecesary columns

movies_rec = movies.drop(["Directors","Stars"],axis=1)
books_rec = books.drop(["Author",],axis=1)
books_rec = books_rec.reindex(columns=["Title","Image","Genre","Description","Rating","Votes"])

We create a column with a identifier

books_rec["Type"] = "Book"
movies_rec["Type"] = "Movie"

We concat both movies and books

recomendation = pd.concat([books_rec,movies_rec])

We create a title identifier, there may be duplicates if a book has a movie adaptation

recomendation["Title_id"] = recomendation["Title"]+"_"+recomendation["Type"]

We filter to have only titles with over 5000 votes

recomendation = recomendation[recomendation["Votes"]>5000]

recomendation.drop_duplicates(inplace=True)
recomendation = recomendation.drop_duplicates(subset="Title_id")
recomendation = recomendation.reset_index()
recomendation = recomendation.drop("index",axis=1)

recomendation

Now we reorder our columns and export to csv

recomendation = recomendation.reindex(columns=["Image","Title","Rating","Votes","Genre","Description","Type","Title_id"])

recomendation.to_csv("../Data/recomendation.csv")
