==> Notebooks/4_Recomendation.ipynb <==
import pandas as pd
import sys
sys.path.append("../")
import src.Resources as src

We load our DataFrame that's clean and ready to be used for the recomendation matrix.

recomendation = pd.read_csv("../Data/recomendation.csv",index_col=0)

recomendation.dropna(inplace=True)

recomendation

We execute the function that prepares the correlation matrix, this function saves two .pkl files that will be used later to do the recomendation, these two pkl ill leave them in a dropbox or drive in github so you can download them.
Anyways you can execute the code and you will generate these two files but it can take a while and you need quite a lot of RAM memmory.

src.matrix(recomendation)
